[SYNPY-1571] Migrate table to new mixin structure (#1179)
* Pull functions out into new mixin model for table/views
