+++ synapseclient/models/__init__.py
@@ -10,7 +10,11 @@ from synapseclient.models.annotations import Annotations
 from synapseclient.models.dataset import Dataset
 from synapseclient.models.file import File, FileHandle
 from synapseclient.models.folder import Folder
-from synapseclient.models.mixins.table_operator import (
+from synapseclient.models.mixins.table_components import QueryMixin
+from synapseclient.models.project import Project
+from synapseclient.models.services import FailureStrategy
+from synapseclient.models.table import Table
+from synapseclient.models.table_components import (
     AppendableRowSetRequest,
     Column,
     ColumnChange,
@@ -21,14 +25,12 @@ from synapseclient.models.mixins.table_operator import (
     JsonSubColumn,
     PartialRow,
     PartialRowSet,
+    QueryResultBundle,
     SchemaStorageStrategy,
     TableSchemaChangeRequest,
     TableUpdateTransaction,
     UploadToTableRequest,
 )
-from synapseclient.models.project import Project
-from synapseclient.models.services import FailureStrategy
-from synapseclient.models.table import Table
 from synapseclient.models.team import Team, TeamMember
 from synapseclient.models.user import UserPreference, UserProfile
 
@@ -58,6 +60,7 @@ __all__ = [
     "ColumnType",
     "FacetType",
     "JsonSubColumn",
+    "QueryResultBundle",
     "query_async",
     "query",
     "query_part_mask_async",
@@ -75,7 +78,8 @@ __all__ = [
 ]
 
 # Static methods to expose as functions
-query_async = Table.query_async
-query = Table.query
-query_part_mask_async = Table.query_part_mask_async
-query_part_mask = Table.query_part_mask
+query_async = QueryMixin.query_async
+# TODO: Static typing on this might cause issues since it's created dynamically
+query = QueryMixin.query
+query_part_mask_async = QueryMixin.query_part_mask_async
+query_part_mask = QueryMixin.query_part_mask
