+++ synapseclient/models/table.py
@@ -13,18 +13,33 @@ from synapseclient.core.constants import concrete_types
 from synapseclient.core.utils import delete_none_keys
 from synapseclient.models import Activity, Annotations
 from synapseclient.models.mixins.access_control import AccessControllable
-from synapseclient.models.mixins.table_operator import (
-    Column,
-    TableOperator,
-    TableRowOperator,
+from synapseclient.models.mixins.table_components import (
+    ColumnMixin,
+    DeleteMixin,
+    GetMixin,
+    QueryMixin,
+    TableBase,
+    TableDeleteRowMixin,
+    TableStoreMixin,
+    TableStoreRowMixin,
+    TableUpsertMixin,
 )
-from synapseclient.models.protocols.table_protocol import TableSynchronousProtocol
+from synapseclient.models.table_components import Column
 
 
 @dataclass()
 @async_to_sync
 class Table(
-    TableSynchronousProtocol, AccessControllable, TableOperator, TableRowOperator
+    AccessControllable,
+    TableBase,
+    TableStoreRowMixin,
+    TableDeleteRowMixin,
+    DeleteMixin,
+    ColumnMixin,
+    GetMixin,
+    QueryMixin,
+    TableUpsertMixin,
+    TableStoreMixin,
 ):
     """A Table represents the metadata of a table.
