+++ tests/integration/synapseclient/models/synchronous/test_dataset.py
@@ -1,89 +1,89 @@
-import pytest
-from synapseclient.models import Dataset, Column, ColumnType
-from synapseclient.core.exceptions import SynapseError
-from typing import List, Dict, Optional
-import pandas as pd
+# from typing import Dict, List
+
+# import pandas as pd
+
+# from synapseclient.models import Dataset
 
 
-def test_create_empty_dataset(syn, project) -> None:
-    """Test creating a new empty dataset"""
-    pass
+# def test_create_empty_dataset(syn, project) -> None:
+#     """Test creating a new empty dataset"""
+#     pass
 
 
-def test_create_dataset_with_files(syn, project, test_files: List[str]) -> None:
-    """Test creating a dataset with one or more files"""
-    pass
+# def test_create_dataset_with_files(syn, project, test_files: List[str]) -> None:
+#     """Test creating a dataset with one or more files"""
+#     pass
 
 
-def test_create_dataset_with_folders(syn, project, test_folders: List[str]) -> None:
-    """Test creating a dataset with one or more folders"""
-    pass
+# def test_create_dataset_with_folders(syn, project, test_folders: List[str]) -> None:
+#     """Test creating a dataset with one or more folders"""
+#     pass
 
 
-def test_create_dataset_with_files_and_folders(
-    syn, project, test_files: List[str], test_folders: List[str]
-) -> None:
-    """Test creating a dataset with both files and folders"""
-    pass
+# def test_create_dataset_with_files_and_folders(
+#     syn, project, test_files: List[str], test_folders: List[str]
+# ) -> None:
+#     """Test creating a dataset with both files and folders"""
+#     pass
 
 
-def test_update_dataset_columns(syn, project, dataset: Dataset) -> None:
-    """Test adding and removing columns from an existing dataset"""
-    pass
+# def test_update_dataset_columns(syn, project, dataset: Dataset) -> None:
+#     """Test adding and removing columns from an existing dataset"""
+#     pass
 
 
-def test_update_dataset_attributes(syn, project, dataset: Dataset) -> None:
-    """Test updating dataset attributes like name, description, annotations"""
-    pass
+# def test_update_dataset_attributes(syn, project, dataset: Dataset) -> None:
+#     """Test updating dataset attributes like name, description, annotations"""
+#     pass
 
 
-def test_update_dataset_add_files_folders(
-    syn, project, dataset: Dataset, files: List[str], folders: List[str]
-) -> None:
-    """Test adding additional files and folders to existing dataset"""
-    pass
+# def test_update_dataset_add_files_folders(
+#     syn, project, dataset: Dataset, files: List[str], folders: List[str]
+# ) -> None:
+#     """Test adding additional files and folders to existing dataset"""
+#     pass
 
 
-def test_update_dataset_upsert_rows(
-    syn, project, dataset: Dataset, rows: List[Dict]
-) -> None:
-    """Test upserting rows into an existing dataset"""
-    pass
+# def test_update_dataset_upsert_rows(
+#     syn, project, dataset: Dataset, rows: List[Dict]
+# ) -> None:
+#     """Test upserting rows into an existing dataset"""
+#     pass
 
 
-def test_update_dataset_remove_rows(
-    syn, project, dataset: Dataset, row_ids: List[str]
-) -> None:
-    """Test removing rows from an existing dataset"""
-    pass
+# def test_update_dataset_remove_rows(
+#     syn, project, dataset: Dataset, row_ids: List[str]
+# ) -> None:
+#     """Test removing rows from an existing dataset"""
+#     pass
 
 
-def test_delete_dataset(syn, project, dataset: Dataset) -> None:
-    """Test deleting an entire dataset"""
-    pass
+# def test_delete_dataset(syn, project, dataset: Dataset) -> None:
+#     """Test deleting an entire dataset"""
+#     pass
 
 
-def test_reorder_dataset_columns(
-    syn, project, dataset: Dataset, column_order: List[str]
-) -> None:
-    """Test reordering columns in a dataset"""
-    pass
+# def test_reorder_dataset_columns(
+#     syn, project, dataset: Dataset, column_order: List[str]
+# ) -> None:
+#     """Test reordering columns in a dataset"""
+#     pass
 
 
-def test_rename_dataset_columns(
-    syn, project, dataset: Dataset, column_renames: Dict[str, str]
-) -> None:
-    """Test renaming columns in a dataset"""
-    pass
+# def test_rename_dataset_columns(
+#     syn, project, dataset: Dataset, column_renames: Dict[str, str]
+# ) -> None:
+#     """Test renaming columns in a dataset"""
+#     pass
 
 
-def test_delete_dataset_columns(
-    syn, project, dataset: Dataset, columns_to_delete: List[str]
-) -> None:
-    """Test deleting columns from a dataset"""
-    pass
+# def test_delete_dataset_columns(
+#     syn, project, dataset: Dataset, columns_to_delete: List[str]
+# ) -> None:
+#     """Test deleting columns from a dataset"""
+#     pass
 
 
-def test_query_dataset(syn, project, dataset: Dataset, query: str) -> pd.DataFrame:
-    """Test querying data from a dataset"""
-    pass
+# def test_query_dataset(syn, project, dataset: Dataset, query: str) -> pd.DataFrame:
+#     """Test querying data from a dataset"""
+#     pass
