+++ tests/integration/synapseclient/models/synchronous/test_table.py
@@ -9,7 +9,7 @@ import pytest
 from pytest_mock import MockerFixture
 
 import synapseclient.models.mixins.asynchronous_job as asynchronous_job_module
-import synapseclient.models.mixins.table_operator as table_module
+import synapseclient.models.mixins.table_components as table_module
 from synapseclient import Evaluation, Synapse
 from synapseclient.core import utils
 from synapseclient.core.constants import concrete_types
